+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 <details>
+<summary>v1.3.0</summary>
+
+## Added
+- Added custom menu for configurations using `ModMenu`
+
+</details>
+
+<details>
 <summary>v1.2.0</summary>
 
 ## Changes